refactor: Removed duplicate implementation of the server command
