+++ src/main/java/ch/heigvd/client/instructions/Exit.java
@@ -1,10 +1,10 @@
-package ch.heigvd.cli.instructions;
+package ch.heigvd.client.instructions;
 
-import ch.heigvd.cli.ClientState;
+import ch.heigvd.client.ClientState;
 
 public class Exit extends Instruction {
     @Override
-    public void exectue(ClientState state) {
+    public void execute(ClientState state) {
         // Noop for now, but could store state for example.
     }
 }
