+++ src/main/java/ch/heigvd/client/instructions/Instruction.java
@@ -1,0 +1,9 @@
+package ch.heigvd.client.instructions;
+
+import ch.heigvd.client.ClientState;
+
+import java.io.IOException;
+
+public abstract class Instruction {
+    public abstract void execute(ClientState state) throws IOException;
+}
